pyproject.toml
[project]
name = "releven-backend"
version = "0.1.0"
description = ""
authors = [{ name = "Kevin Stadler", email = "kevin.stadler@oeaw.ac.at" }]
requires-python = "~=3.12"
dependencies = [
    "rdfproxy>=0.6.0",
    "fastapi[standard]>=0.115.4,<0.116",
    "gitpython>=3.1.43,<4",
]

[tool.uv.sources]
rdfproxy = { git = "https://github.com/acdh-oeaw/rdfproxy" }

